[INJIMOB-3257]: Refactor Hardcoded data
